Final_py1
It is a program that contains 2 web pages. First, it works for logining user and if a database has a date about the user it will give and save token in a database and redirect them to the next page. Second, it works for searching and summarizing information about concrete token. It uses a bootstrap on each HTML page.